Work on project. Stage 1/5: Unfair computer
escription
Rock, paper, scissors is a well-known hand game. Each one of two players simultaneously forms one of three shapes with their hands, and then, depending on the chosen shapes, the winner is determined: rock beats scissors, paper wins over rock, scissors beat paper.
The game is widely used to make a fair decision between equal options.

So, let's start with an unfair version! :)

Write a program that reads input specifying which option the user has chosen. Then your program should make the user lose! That is, your program should always choose an option that defeats the one picked by the user. After finding this option, output a line Sorry, but computer chose <option>, where <option> is the name of option that the program picked depending on the user's input.
For example, if the user enters rock, the program should print Sorry, but computer chose paper and so on.

Objectives
Your program should:

Take an input from the user
Find an option that wins over the user's option
Output a line: Sorry, but computer chose <option>